+++ src/test/java/eu/kidf/diversicon/cli/test/DiverCliIT.java
@@ -4,27 +4,17 @@ import static org.junit.Assert.*;
 
 import java.io.File;
 import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-
-import org.ini4j.Wini;
 import org.junit.Test;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import eu.kidf.diversicon.cli.DiverCli;
 import eu.kidf.diversicon.cli.commands.InitCommand;
-import eu.kidf.diversicon.cli.commands.LogCommand;
 import eu.kidf.diversicon.core.Diversicon;
-import eu.kidf.diversicon.core.internal.Internals;
 import eu.kidf.diversicon.data.DivWn31;
 
 /**
  * @since 0.1.0
  */
-public class DiverCliIT extends DiverCliTestBase {
-
-    private static final Logger LOG = LoggerFactory.getLogger(DiverCliIT.class);
+public class DiverCliIT extends DiverCliTestBase {  
 
     /**
      * @since 0.1.0