+++ src/test/java/eu/kidf/diversicon/cli/test/DiverCliTestBase.java
@@ -1,13 +1,10 @@
 package eu.kidf.diversicon.cli.test;
 
-import java.io.File;
 import java.io.IOException;
-import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 
-import org.apache.commons.io.FileUtils;
 import org.junit.After;
 import org.junit.Before;