+++ src/test/java/eu/kidf/diversicon/cli/test/DocsGenIT.java
@@ -49,6 +49,7 @@ import eu.kidf.diversicon.cli.exceptions.DiverCliIoException;
 import eu.kidf.diversicon.core.BuildInfo;
 import eu.kidf.diversicon.core.ImportJob;
 import eu.kidf.diversicon.core.exceptions.InvalidImportException;
+import eu.kidf.diversicon.core.exceptions.InvalidXmlException;
 import eu.kidf.diversicon.data.DivWn31;
 import eu.kidf.diversicon.data.Examplicon;
 import eu.kidf.diversicon.data.Smartphones;
@@ -133,11 +134,6 @@ public class DocsGenIT extends DiverCliTestBase {
      */
     private static final String MYPRJ = "myprj";
     
-    /**
-     * @since 0.1.0
-     */
-    private static final String WN31 = "wn31";
-
     /**
      * @since 0.1.0
      */
@@ -478,8 +474,8 @@ public class DocsGenIT extends DiverCliTestBase {
             String cap = stopCaptureSlf4j();
 
             String val = "```bash\n"
-                    + "> " + DIVERCLI + " " + sbStr.toString() + "\n"
-                    + cap.replace(System.getProperty(DiverCli.SYSTEM_PROPERTY_WORKING_DIR) + "/", "")
+                    + "> " + DIVERCLI + " " + (sbStr.toString() + "\n"
+                    + cap).replace(System.getProperty(DiverCli.SYSTEM_PROPERTY_WORKING_DIR) + "/", "")
                          .replace(System.getProperty(DiverCli.SYSTEM_PROPERTY_WORKING_DIR), "")
                          .replace(System.getProperty(DiverCli.SYSTEM_PROPERTY_USER_HOME) + "/", "/home/divergeek/")
                          .replace(System.getProperty(DiverCli.SYSTEM_PROPERTY_USER_HOME), "/home/divergeek/")                         
@@ -528,10 +524,11 @@ public class DocsGenIT extends DiverCliTestBase {
         emptyInit();
         try {
             diver("smartphones.import.failed",            
-                ImportXmlCommand.CMD, "--author", "\"John Doe\"", "--description", "Some test import",
+                ImportXmlCommand.CMD, "--author", "\"John Doe\"", "--description", "\"Some test import\"",
                  Smartphones.XML_URI );
+            Assert.fail("Shouldn't arrive here!");
         } catch (InvalidImportException ex){
-            
+            LOG.debug("Got expected exception: ", ex);
         }
     }    
 
@@ -543,7 +540,7 @@ public class DocsGenIT extends DiverCliTestBase {
         wn31Init();
         diver("smartphones.import.success",            
                 ImportXmlCommand.CMD,
-                "--author", "\"John Doe\"", "--description", "Some test import",
+                "--author", "\"John Doe\"", "--description", "\"Some test import\"",
                  Smartphones.XML_URI );
         diver("smartphones.import.success.log",            
                 LogCommand.CMD);
@@ -558,7 +555,7 @@ public class DocsGenIT extends DiverCliTestBase {
         wn31Init();
         diver("smartphones.examplicon.import.success",            
                 ImportXmlCommand.CMD,
-                "--author", "\"John Doe\"", "--description", "Some test import",
+                "--author", "\"John Doe\"", "--description", "\"Some test import\"",
                  Smartphones.XML_URI,
                  Examplicon.XML_URI);
     }        
@@ -571,12 +568,12 @@ public class DocsGenIT extends DiverCliTestBase {
         wn31Init();
         diver("smartphones.import.skipaugment",            
                 ImportXmlCommand.CMD,
-                "--skip-augment", "--author", "\"John Doe\"", "--description", "Some test import",
+                "--skip-augment", "--author", "\"John Doe\"", "--description", "\"Some test import\"",
                  Smartphones.XML_URI);
         diver("examplicon.import.skipaugment",            
                 ImportXmlCommand.CMD,
-                "--skip-augment", "--author", "\"John Doe\"", "--description", "Some test import",
-                 Smartphones.XML_URI);
+                "--skip-augment", "--author", "\"John Doe\"", "--description", "\"Some test import\"",
+                 Examplicon.XML_URI);
         diver("smartphones.examplicon.dbaugment",            
                 DbAugmentCommand.CMD);        
     }    
@@ -590,11 +587,13 @@ public class DocsGenIT extends DiverCliTestBase {
         try {
             diver("badexamplicon.import",            
                     ImportXmlCommand.CMD,
-                    "--author", "\"John Doe\"", "--description", "Some test import",
+                    "--author", "\"John Doe\"", "--description", "\"Some test import\"",
                      BAD_EXAMPLICON_URL );
+            Assert.fail("Shouldn't arrive here!");
         } catch (InvalidImportException ex){
-            
+            LOG.debug("Got expected exception: ", ex);
         }
+            
     }    
     
         
@@ -608,7 +607,7 @@ public class DocsGenIT extends DiverCliTestBase {
         diver("smartphones.import.force",            
                 ImportXmlCommand.CMD,
                 "--force",
-                "--author", "\"John Doe\"", "--description", "Some test import",
+                "--author", "\"John Doe\"", "--description", "\"Some test import\"",
                  Smartphones.XML_URI );          
     }    
     
@@ -629,9 +628,14 @@ public class DocsGenIT extends DiverCliTestBase {
     @Test
     public void badExampliconValidate() {       
         emptyInit();
-        diver("badexamplicon.validate",            
-                ValidateCommand.CMD,
-                BAD_EXAMPLICON_URL);          
+        try {
+            diver("badexamplicon.validate",            
+                    ValidateCommand.CMD,
+                    BAD_EXAMPLICON_URL);
+            Assert.fail("Shouldn't arrive here!");
+        } catch (InvalidXmlException ex){
+            LOG.debug("Got expected exception. ", ex);            
+        }
     }
     
     /**
@@ -642,8 +646,8 @@ public class DocsGenIT extends DiverCliTestBase {
         emptyInit();
         diver("divupper.export",            
                 ExportXmlCommand.CMD,
-                "--name", "acme-upper-lexres",
-                "acme-upper-lexres.xml" ); 
+                "--name", "div-upper",
+                System.getProperty(DiverCli.SYSTEM_PROPERTY_WORKING_DIR) + "acme-upper-lexres.xml" ); 
     }    
     
     /**